requirements.txt
numpy>=1.20.0
pandas>=1.3.0
spacy>=3.0.0
nltk>=3.6.0
transformers>=4.18.0
pyyaml>=6.0
tqdm>=4.62.0
tensorflow>=2.15.0
wheel>=0.45.0
scikit-learn>=1.0.0
sentence-transformers>=2.0.0
scipy>=1.7.0
matplotlib>=3.5.0
pypdf>=3.0.0
pdfplumber>=0.11.0
tabulate>=0.9.0
langchain-google-genai>=0.1.0
langchain-core>=0.1.0
huggingface_hub>=0.20.0
faiss-cpu>=1.7.0
seaborn>=0.12.0
python-dotenv>=0.19.0
langchain>=0.1.0